requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
PyJWT==2.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7  # For PostgreSQL (optional)
SQLAlchemy==2.0.20
bcrypt==4.0.1